+++ src/utils/index.ts
@@ -4,4 +4,4 @@ export * from "./matcher";
 export * from "./others";
 export * from "./packages";
 export * from "./user";
-export * from "./validater";
+export * from "./validator";
